Suggest adding a test fixture for tuple
